Add files via upload
Added 'Get_Match_Data.ipynb', an iPython notebook which accesses the RIOT games API to aggregate match data for League of Legends ranked queues. The output of the file is the 'ranked_matches.csv' file included in this folder, as well as two pickle files which act as checkpoints in the python notebook (the API requests are rate limited, so getting a lot of data can be a time consuming process).